Support for staging App Store storefront region override on iOS (#424)
Task/Issue URL:
https://app.asana.com/0/414709148257752/1209352855429093/f
cc: @federicocappelli 

### Description
Similar to how it is done on macOS implement debug menu option for
setting a fixed storefront region override that can be used during
testing. Set override region is used instead of getting value via
StoreKit API to determine available subscription products for purchase.
override value can be set to one of three values: "None", "USA", "Rest
of World"
override is only used when internal user mode is enabled and using
staging environment

Implementation for using override is already in place (used by macOS),
we miss debug menu option to set the value and configuration to read and
pass the current override value to the StorePurchaseManager.

⚠️ Change of override may require app restart to take effect. 

### Testing Steps (iOS)
Ensure you are using Alpha build for testing and test on real device

**When:**
- Internal user mode is enabled
- subscription environment is set to staging
- region override is set to "None" 

**Then:** 
- Purchase page shows products (description on the page, purchase
options in the footer and matching ToS) for the current Apple IDs region
(or one shown in the debug menu)

---

**When:**
- Internal user mode is enabled
- subscription environment is set to staging
- region override is set to "USA" 

**Then:** 
- Purchase page shows products (description on the page, purchase
options in the footer and matching ToS) for the US region
- After purchase user is granted 3 US Privacy Pro entitlements (VPN,
PIR, ITR) and its features are listed in Settings
 
---

**When:**
- Internal user mode is enabled
- subscription environment is set to staging
- region override is set to "Rest of World"

**Then:** 
- Purchase page shows products (description on the page, purchase
options in the footer and matching ToS) for the US region
- After purchase user is granted 2 Rest of World Privacy Pro
entitlements (VPN, ITR) and its features are listed in Settings

---

**When:**
- Internal user mode is disabled
- subscription environment is set to staging
- region override is set to "USA" or "Rest of World" 

**Then:** 
- Override is not applied and purchase page shows products (description
on the page, purchase options in the footer and matching ToS) for the
current Apple IDs region (or one shown in the debug menu)

---

**When:**
- Internal user mode is enabled
- subscription environment is set to production
- region override is set to "USA" or "Rest of World" 

**Then:** 
- Override is not applied and purchase page shows products (description
on the page, purchase options in the footer and matching ToS) for the
current Apple IDs region (or one shown in the debug menu)


### Impact and Risks
None: Internal tooling, documentation

#### What could go wrong?
Little to no risks, region override is only available in the debug menu
and its application is guarded by internal user mode and staging
environment.
Also implementation to apply the logic is already present as being
already used on macOS.

---
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)