Fix macos sync UI tests (#513)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1201037661562251/task/1209998672880350?focus=true

### Description

There's been a slight change in the sync logic determining what is a
recovery. Previously, the recovery codes was used for connecting as well
as recovery, but now it's only used for recovery. This means that the
recovery code modal will not show when logging in in the UI tests (as
they only ever use a recovery code).

This change should capture both the old and new flow.