improvements in subscription with AuthV2 resiliency (#493)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1205842942115003/task/1209974732411306
CC: @miasma13 

### Description

This fixes a bug I noticed in iOS that causes a couple of issues when
the device is connected to a wifi with 100% packet loss, for example, a
wifi with a captive portal before authentication.

How to reproduce the bugs:
- Start the app
- Enable AuthV2 feature flag
- Kill the app
- Start the app and subscribe/restore the subscription
- Enable 100% packet loss in the iOS network link conditioner
- Kill the app
- Start the app
- Bug #1: The app needs a long time to start (60s, the API request
timeout), this was due to a blocking semaphore in the token provider
- Bug #2: remove the subscription from settings: nothing happens for
60s, the logout API call was blocking the operation.

Fixes:
- iOS token provider is now async; the synchronous code has been moved.
- The logout now runs the logout API call in a Task

### Known bugs

Both with AuthV1 and V2, when the app is opened for the first time
without an internet connection, the subscription setting is not
displayed.

### Testing Steps

1. Test the iOS app's standard operations while enabling 100% packet
loss in the iOS network link conditioner
2. Test the macOS app's standard operations while enabling 100% packet
loss in the iOS network link conditioner

---
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)