Display an info icon to notify users about a crashed tab (#550)
Task/Issue URL: https://app.asana.com/1/137249556945/project/72649045549333/task/1209977265286175?focus=true

Description:
This change displays an info icon next to the tab favicon for 20 seconds after a single crash.
The icon can be activated and it would show a popover explaining that the tab has crashed.
The icon is displayed on both active and inactive tabs when they crash.
If the popover is open, the 20s timer is invalidated. If the user clicks anywhere when the popover
is open, the popover gets closed and icon disappears.
The crash icons replaces audio icon on the tab bar view item.