WIP - Native chat box for duck.ai  (#698)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1204167627774280/task/1210195143126427?focus=true

### Description
Initial set of changes required to add the native input field for
duck.ai.
This introduces a new dependency on AIChatViewController, allowing it to
accept any SwiftUI view that will be laid out at the bottom of the
screen where the chat box should be.
It also adds handling for new FE methods that are not yet in production