[macOS] Allow smaller minimum window size (#656)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1209825025475019/task/1210159114504360?focus=true

### Description

This is the final PR to merge these changes:

* https://github.com/duckduckgo/apple-browsers/pull/536
* https://github.com/duckduckgo/apple-browsers/pull/577
* https://github.com/duckduckgo/apple-browsers/pull/648

Additional changes in this PR:

* Resolves merge conflicts:
* Reverts earlier change that moved the Dax logo inside the address bar
(this is feature flagged in `main` as part of the ongoing visual
refresh).
* Now supports resizing the window to the smallest width with the Dax
logo inside or outside the address bar.
* The address bar stays the same width when activating the text field.

Not resolved:

From
https://github.com/duckduckgo/apple-browsers/pull/648#pullrequestreview-2814280977:

> Address bar string is truncated too early, even with the hidden
bookmarks button, there’s still room for more text.

This happens because the passive text field is moved to the left, to
accommodate the AI chat button:


https://github.com/duckduckgo/apple-browsers/blob/28ea3a57332a19661f8edca08aa47113e84b1b6c/macOS/DuckDuckGo/NavigationBar/View/AddressBarViewController.swift#L670

As far as I can see, it only happens when the AI chat button is visible
_and_ the zoom/permissions buttons are shown. Because of how the layout
is currently done, I can’t find a reasonable way to adjust it right now.

### Testing Steps

Test with the `visualRefresh` feature flag enabled and disabled:

1. Confirm the window can be resized down to 544 pt; opening new tabs
and windows retains that width.
2. All settings pages can be resized, viewed, and interacted with at the
narrow width.
3. The bookmarks management page can be resized, viewed, and interacted
with at the narrow width.
4. The Home button and shortcuts (Passwords, Bookmarks, Downloads, VPN)
can be enabled from the View menu. When needed, these buttons will move
to an overflow menu to fit the narrow width.

### Impact and Risks
Low: Minor visual changes

#### What could go wrong?
The window resize not working as expected. In particular, the resize not
working correctly when the Dax logo appears outside the address bar (on
the home page).

---
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)