More flakey sync test fixes (#934)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1205237866452338/task/1210451347360805?focus=true

Description:
These test were flaking out. They worked on my machine initially, also worked by fluke on CI on my PR, then stared failing on main.

I initially merged another PR which attempted to reduce the number of async waits. I even continued to think that was where the issue was here and stopped using some helpers which depended on XCTExpectations. However, it hit me that in fact the reason was likely that we're trying to compare two base64 code generated from a serialised SyncCode.RecoveryKey when in fact, they could differ because of variations in the order of the JSON KV pairs. This passed on the first try so I'm hoping that this resolves the issue