Add dynamic “Complete Your Setup” section to Settings screen (#1026)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/72649045549333/task/1210410396636449?focus=true
Tech Design URL:
CC:

### Description
Adds new section to the top of the Settings screen with the options to
set the browser as default and to import passwords (if certain
conditions are met). This is behind a feature flag and may only be a
short-lived change, depending on metrics.