[iOS][Set As Default] Add Modal prompts (#1110)
Task/Issue URL: https://app.asana.com/1/137249556945/project/72649045549333/task/1210454186090914
Tech Design URL:
CC:

### Description

This PR adds the UI for the modal prompt to set the browser as default.
I have extracted the MetricBuilder from onboarding and put as standalone package in iOS local packages (I think it makes sense to have it here since it’s iOS only).
I will make a PR to replace the version in Onboarding with the one from
the package later on.

### Impact and Risks
Low, the feature is not plugged in.