Fix an issue where SAD/ATT prompts would not show if user do not close the app after onboarding (#1154)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1206329551987282/task/1210646559468282?focus=true
Tech Design URL:
CC:

### Description

This PR addresses an issue with the onboarding completion check. The
value used to determine if onboarding was completed was passed as a
variable rather than utilizing a closure that evaluates the condition on
demand.

**Issue:**

The prompt to the user was not displayed after they finished onboarding
unless the app was closed in the meantime.


**Solution:**

By implementing a closure for the onboarding completion check, the app
will now evaluate the status each time, ensuring that the prompt is
presented immediately after onboarding is completed, regardless of
whether the app was closed or not.

### Testing Steps
Ensure new unit test pass.

### Impact and Risks
Low: Improvement on existing code.