[iOS][Set As Default Browser] Pixels implementation (#1226)
Task/Issue URL: https://app.asana.com/1/137249556945/project/72649045549333/task/1210577908775554
Tech Design URL:
CC:

### Description

This PR adds pixels implementation for the Default Browsers Prompt. 

### Testing Steps

To facilitate testing:
- Comment out `guard let prompt = promptTypeDecider.promptType() else {
return nil }` in ‘DefaultBrowserPromptCoordinator’ Line 86
- Change `return DefaultBrowserPromptPresentationType(prompt)` to return
DefaultBrowserPromptPresentationType(.firstModal) in
`DefaultBrowserPromptCoordinator` Line 91.

1. When the modal appear ensure that
`m_set-as-default-prompt_modal-shown` is fired. Ensure that the
‘numberOfModalShown' parameter reflects the number of modal shown.
2. Send the app to the background and foreground 10 times and ensure
that `numberOfModalShown` increase from 1 to 10.
3. Tap 'Set As Default Browser’ button and ensure that
m_set-as-default-prompt_modal-set-as-default-browser-button-action` is
fired and numberOfModalShown parameter reflect 10.
4. Go Back to the app.
5. The modal should show again.
6. Ensure that `m_set-as-default-prompt_modal-shown` is fired but this
time ‘numberOfModalShown' parameter is ‘10+’
7. Tap 'Set As Default Browser’ button and ensure that
`m_set-as-default-prompt_modal-set-as-default-browser-button-action` is
fired and numberOfModalShown parameter reflect 10+.
8. Tap close button. ensure that when ‘Close’ button is tapped then
`m_set-as-default-prompt_modal-closed-button-action` is fired.
9. Show modal again.
5. Tap ‘Don’t ask again’ button. Ensure that
`m_set-as-default-prompt_modal-do-not-ask-again-button-action` is fired.

### Impact and Risks
Wrong monitoring reported

#### What could go wrong?
I implemented unit tests for pixels to measure engagement to validate
the behaviour.