macOS: Convert Freemium PIR from Experiment to Full Feature (PR # 2)  (#1166)
Task/Issue URL:
https://app.asana.com/1/137249556945/task/1210613732778387
Tech Design URL:
https://app.asana.com/1/137249556945/project/481882893211075/task/1209397034199086

### Description
This PR (2 of 2) implements the remaining changes to convert Freemium
PIR from an experiment to a full feature. Changes here are:
1. Add `isUserEligibleForFreeTrial` property to `DBPUIHandshakeUserData`
model and add plumbing to set the value. (**NOTE**: This can't be
manually tested as it’s not used on the frontend yet, but we can rely on
unit tests. Also by testing Freemium PIR works as expected, this is
confirming that this property is just additive, i.e not breaking
anything)
2. Add copy translations.