Match NTP favorite style with unified input overlay (#1381)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1206226850447395/task/1210742505823399?focus=true
Tech Design URL:
CC:

### Description

This PR aligns the design of the NTP favorites and the favorites overlay
shown in the focus mode. For parts of the alignment, an additional
variable needed to be introduced to not modify production state.

* The way how icons are being rendered on NTP was fixed - previously
they were unnecessarily scaled up when size was smaller than the
bounding box.
* The reload done after `viewDidLayoutSubviews` on FavoritesOverlay was
removed. It proved to be redundant, but was causing the flicker during
the transition.