Open summarization source links in new tab when Duck.ai chat is in a full-size tab (#1452)
Task/Issue URL: https://app.asana.com/1/137249556945/task/1210855725388312?focus=true

Description:
This change changes behavior of the summarization source link when in a full-size tab.
The link will now open in a new tab instead of the current tab, and when such tab is already open,
it will switch to the open tab instead.
The behavior for sidebar chats is the same and clicking the link will open the website in the current tab.