[iOS] Remove DefaultBrowserManager and Service from the iOS project and use the one defined in local Swift Package (#1392)
Task/Issue URL: https://app.asana.com/1/137249556945/task/1210636953633415
Tech Design URL:
CC:

### Description

This PR removes the manager and service used to check if the browser is the default defined in the iOS project.
These classes had previously been moved in a Package but we did not remove it yet as we were running an experiment. Now that the experiment is completed we can safely remove the duplicate code.

### Testing Steps
1. Ensure CI tests pass.

Smoke Test (Test on a real device)
1. Remove any previous builds installed.
2. Ensure that DDG is not set as the default browser.
3. Complete onboarding.
4. Go to system settings and change the date to a couple of days in the
future.
5. Foreground the app.
6. Ensure a modal promo is prompted asking to to set DDG as default
browser.

### Impact and Risks
Low: The changes to the onboarding flow are safe as the experiment is
not currently running.

#### What could go wrong?
The DefaultBrowserInfoStore is the only class that changed. This was
mapping between two identical type. Now the compiler ensures safety and
unit test covers that it still works.

### Quality Considerations
<!-- 
Focus on what matters for your changes:
- What edge cases exist?
- How does this affect performance?
- What monitoring have you added?
- What documentation needs updating?
- How does this impact privacy/security?
-->

### Notes to Reviewer
<!-- Anything specific you want reviewers to focus on -->

---
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)