macOS: Improve Quality of Direct User Feedback (#1610)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1204006570077678/task/1210990797327013?focus=true
Tech Design URL:
https://app.asana.com/1/137249556945/project/481882893211075/task/1210859240378092
CC:

### Description
We are changing how we ask for feedback from our users. Instead of
showing a dialog with some options and then the input text box, we will
display different problems/new features and let them select the ones
they are interested in, providing more text if necessary. The reason is
that we want more qualitative feedback, and we are not getting that with
the current flow.

**Technical**
- I’ve taken the opportunity to also remove the `FEEDBACK` preprocessor
flag from the code, given that is not currently being used (we have it
in all builds)
- There is a Technical Design on how we will send the data to the
backend. TLDR; we will use the `sub-category` parameter from the
backend, and send the different selected options as Asana tags, so then
we can easily filter and query the feedback.

**Demo**


https://github.com/user-attachments/assets/ca5ffd65-4176-4cb7-a0b5-e46f1c83b7c3



### Testing Steps
1. Open the More Options menu -> Send Feedback
2. From there, select the Report a Browser Problem or Request New
Feature options
3. Play with the form
4. Submit your feedback. NOTE: The feedback is working, so the feedback
you submit will end in the `macOS Feedback` project, make sure to delete
it after testing.

### Impact and Risks
Medium: Could disrupt specific features or user flows

#### What could go wrong?
We are aware that this could have an impact on feedback quantity, but at
the moment, we are prioritizing quality over quantity.

### Quality Considerations
Nothing in particular.

### Notes to Reviewer
Nothing in particular.

—
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)