Enables the Bitwarden Web Extension for internal users (#1823)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1206580121312550/task/1211182965070662?focus=true

### Description

Enables the Bitwarden Web Extension for internal users

### Testing Steps

#### Feature flag OFF

1. Turn the "webExtensions" feature flag off.
2. Ensure you're an internal user.
3. Restart the app.
4. Run the app, make sure password managers in both DMG and Sparkle
builds work as usual.

#### Feature flag ON

Test in both DMG and App Store.

1. Turn the "webExtensions" feature flag on.
2. Ensure you're an internal user.
3. Restart the app.
4. Run the app - install extensions through the debug menu and make sure
they work reasonably well.
5. Uninstall all extensions.
6. Go to App Settings > Passwords & Autofill > Select "Bitwarden
Extension"
7. Make sure the Bitwarden extension works

### Impact and Risks

Low: only risk is if any of the code when the feature flag is OFF has
been changed unintentionally.

#### What could go wrong?

Mostly focus on ensuring the code with the feature flag OFF is the same
as it was before.
If we're affecting internal users with issues we can tackle as bugs, so
it's slightly less problematic.

### Quality Considerations

NA

### Notes to Reviewer

Apologies for the large-ish PR: there's a lot of noise coming from
moving the build configuration to feature flag, and other small
adjustments across files.

This PR should be fine as long as we ensure the code with the feature
flag OFF works as usual. Please keep this in mind during review.

---
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)

---------

Co-authored-by: Tomas Strba <tom@duckduckgo.com>